# src/swaggerConfig.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cinema API",
    "version": "1.0.0",
    "description": "Cinema Api in TS"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/rooms": {
      "get": {
        "summary": "List all rooms",
        "description": "Retrieve a list of rooms with pagination and sorting options.",
        "operationId": "listRooms",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of rooms to return",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number to retrieve",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "ascending",
            "in": "query",
            "description": "Sort order",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Field to order the rooms by",
            "required": false,
            "schema": {
              "type": "string",
              "default": "id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of rooms",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Room"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/rooms/{id}/shows": {
      "get": {
        "summary": "Get Shows of a room",
        "description": "Retrieve shows of a room by its id.",
        "operationId": "getShowsRoom",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Id of the asked room",
            "required": true,
            "schema": {
              "type": "number",
              "default": "id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of shows of the movie",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "items": {
                        "$ref": "#/components/schemas/Movie"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "404": {
            "description": "Unknown Room"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/rooms/{id}": {
      "post": {
        "summary": "Create a room",
        "description": "Create a room",
        "operationId": "createRoom",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Name of the room",
            "required": true,
            "schema": {
              "type": "string",
              "default": 10
            }
          },
          {
            "name": "description",
            "in": "query",
            "description": "Description of the room",
            "required": true,
            "schema": {
              "type": "string",
              "default": 10
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Type of the room",
            "required": true,
            "schema": {
              "type": "string",
              "default": 10
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "State of the room",
            "required": true,
            "schema": {
              "type": "boolean",
              "default": 10
            }
          },
          {
            "name": "handicapAvailable",
            "in": "query",
            "description": "Is handicapAvalaible of the room",
            "required": true,
            "schema": {
              "type": "boolean",
              "default": 10
            }
          },
          {
            "name": "capacity",
            "in": "query",
            "description": "Capacity of the room",
            "required": true,
            "schema": {
              "type": "number",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A room",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "items": {
                        "$ref": "#/components/schemas/Room"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "404": {
            "description": "Unknown Room"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "duration": {
            "type": "integer",
            "format": "int64",
            "description": "Duration in minutes"
          }
        }
      },
      "Room": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "state": {
            "type": "boolean"
          },
          "handicapAvailable": {
            "type": "boolean"
          },
          "capacity":  {
            "type": "number"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}